Mark RelocatorRemapper as internal
